pyproject.toml
[project]
name = "ai-business-automation-demo"
version = "1.0.0"
description = "AI Business Automation Demo Platform - Showcasing AI agents automating real business workflows"
authors = [
    {name = "Kyle M", email = "kyle@example.com"}
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.18",

    # Data Processing
    "pandas>=2.1.0",
    "numpy>=1.26.0",

    # Visualization
    "plotly>=5.18.0",

    # Machine Learning / Forecasting
    "prophet>=1.1.5",

    # AI/LLM Libraries
    "google-generativeai>=0.8.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=0.2.53",

    # Environment Management
    "python-dotenv>=1.0.0",

    # Embedding and Model Dependencies
    "sentence-transformers>=2.7.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.38.0",
    "tokenizers>=0.15.0",
    "huggingface-hub>=0.20.0",
    "safetensors>=0.4.2",

    # RAG Chatbot Dependencies
    "chromadb>=0.4.22",
    "duckduckgo-search>=4.1.0",
    "pypdf>=4.0.0",
    "tiktoken==0.8.0",
    "arxiv>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "black==23.12.1",
    "flake8==7.0.0",
    "mypy==1.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agents", "utils"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | data
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
